restructured logging
